arrumando atualizacao de caixa,e adicionando o gerarjson na nfe

+++ pdv/uintegracaoodoo.pas
@@ -37,34 +37,63 @@ var k: Integer;
   dadosJson: TJsonNode;
   responseData: String;
   listaArquivos : TStringList;
+  //logs:TextFile;
+  ver : String;
 begin
+  // log
+  ver := 'C:\home\log_integracao.txt';
+  //AssignFile(logs, ver);
   // VER SE EXISTE ARQUIVO
   listaArquivos := TStringList.Create;
   try
+    //if FileExists(ver) then
+    //  Append(logs)
+    //else
+    //  Rewrite(logs);
+
     FindAllFiles(listaArquivos, path_integracao, '*.txt', true);
     for k:=0 to Pred(listaArquivos.Count) do
     begin
+      //Writeln(logs, 'Iniciando integracao : ' + FormatDateTime('mm/dd/yyyy hh:MM', Now));
+      //Writeln(logs, 'Path : ' + path_integracao);
+
       // SE EXISTE ENVIA
       postJson := TJSONObject.Create;
       dadosJson := TJsonNode.Create;
       postJson.Add('title', 'Enviando Movimento');
+
+      //Writeln(logs, 'Empresa : ' + nome_empresa_integra);
+
       postJson.Add('body', nome_empresa_integra);
       dadosJson.LoadFromFile(listaArquivos[k]);
+      ver := listaArquivos[k];
+      //Writeln(logs, 'Arquivo : ' + listaArquivos[k]);
+
       postJson.Add('tab_venda', dadosJson.ToString);
       postJson.Add('userId', 1);
       With TFPHttpClient.Create(Nil) do
+      begin
         try
           AddHeader('Content-Type', 'application/json');
           RequestBody := TStringStream.Create(postJson.AsJSON);
           responseData := Post(path_integracao_url);
-          if FileExists(path_integracao + responseData + '_' + cod_caixa_integra + '.txt') then
-            DeleteFile(path_integracao + responseData + '_' + cod_caixa_integra + '.txt');
+          if (responseData <> 'N') then
+          begin
+            //Writeln(logs, 'Existe arquivo : ' + path_integracao + responseData + '_' + cod_caixa_integra + '.txt');
+            if FileExists(path_integracao + responseData + '_' + cod_caixa_integra + '.txt') then
+            begin
+              //Writeln(logs, 'Excluindo arquivo : ' + path_integracao + responseData + '_' + cod_caixa_integra + '.txt');
+              DeleteFile(path_integracao + responseData + '_' + cod_caixa_integra + '.txt');
+            end;
+          end;
         finally
          Free;
         end;
+      end;
     end;
   finally
     listaArquivos.Free;
+    //CloseFile(logs);
   end;
 end;
